feat: add initial structure for the model repository with e_coli_core
Biochemical network reconstructions have become popular tools in systems
biology. Metabolicnetwork reconstructions are biochemically, genetically, and
genomically (BiGG) structured databases of biochemical reactions and
metabolites. They contain information such as exact reaction stoichiometry,
reaction reversibility, and the relationships between genes, proteins, and
reactions. Network reconstructions have been used exten- sively to study the
phenotypic behavior of wild-type and mutant stains under a variety of
conditions, linking genotypes with phenotypes. Such phenotypic simulations have
allowed for the prediction of growth after genetic manipulations, prediction
of growth phenotypes after adaptive evolution, and prediction of essential
genes. Additionally, because network reconstructions are organism specific, they
can be used to understand differences between organisms of species in a
functional context. There are different types of reconstructions representing
various types of biological networks (metabolic, regulatory, transcription/
translation). This chapter serves as an introduction to metabolic and
regulatory network reconstructions and models and gives a complete description
of the core Escherichia coli metabolic model. This model can be analyzed in any
computational format (such as MATLAB or Mathematica) based on the information
given in this chapter. The core E. coli model is a small-scale model that can be
used for educational purposes. It is meant to be used by senior undergraduate
and first-year graduate students learning about constraint-based modeling and
systems biology. This model has enough reactions and pathways to enable
interesting and insightful calculations, but it is also simple enough that the
results of such calculations can be understood easily.