feat: Change iJR904 to iAF1260
An updated genome-scale reconstruction of the metabolic network in Escherichia
coli K-12 MG1655 is presented. This updated metabolic reconstruction includes:
(1) an alignment with the latest genome annotation and the metabolic content of
EcoCyc leading to the inclusion of the activities of 1260 ORFs,
(2) characterization and quantification of the biomass components and
maintenance requirements associated with growth of E. coli and
(3) thermodynamic information for the included chemical reactions. The
conversion of this metabolic network reconstruction into an in silico model is
detailed. A new step in the metabolic reconstruction process, termed
thermodynamic consistency analysis, is introduced, in which reactions were
checked for consistency with thermodynamic reversibility estimates.
Applications demonstrating the capabilities of the genome-scale metabolic model
to predict high-throughput experimental growth and gene deletion phenotypic
screens are presented. The increased scope and computational capability using
this new reconstruction is expected to broaden the spectrum of both basic
biology and applied systems biology studies of E. coli metabolism.